Added -std=c++11 to fix build errors on Linux

+++ sdk/wxshapeframework/src/wxxmlserializer/PropertyIO.cpp
@@ -162,7 +162,7 @@ wxString xsDoublePropIO::ToString(const double& value)
 {
     wxString sVal;
 
-    if( wxIsNaN(value) )
+    if( wxIsNaN((long double)value) )
     {
         sVal = wxT("NAN");
     }
